fix bugs

+++ __main__.py
@@ -75,36 +75,42 @@ local  :- To Where | gdrive url
         file_ext = file.split('.')[-1]
         file_name = file.replace(self.remote, '')
 
-        orig_size, comp_size, status = is_incomplete(file.replace(self.remote, self.local), file)
-
-        if status:
-            print(saveas, ' not compressed completely')
-            print(f'orig Size {orig_size//1024**2}MB : compressed Size {comp_size//1024**2}MB')
-            print('removing file')
-            os.unlink(file.replace(self.remote, self.local))
+        # if exists check for size
+        if os.path.exists(file.replace(self.remote, self.local)):
+            orig_size, comp_size, status = is_incomplete(file.replace(self.remote, self.local), file)
+            if status:  # incomplete
+                print(saveas, ' not compressed completely')
+                print(f'orig Size {orig_size//1024**2}MB : compressed Size {comp_size//1024**2}MB')
+                print('removing file')
+                os.unlink(file.replace(self.remote, self.local))
 
         if self.should(saveas):
-            ffmpeg_cmd = "ffmpeg -i " + self.valid_unix_name(file) + "\
-                    -b:a 64k -ac 1 -vf scale=\"'w=-2:h=trunc(min(ih," + str(self.res) + ")/2)*2'\" \
-                    -crf 32 -profile:v baseline -level 3.0 -preset slow -v error -strict -2 -stats \
-                    -y -r 20 " + saveas
-            print('compressing\t', file_name)
-            os.system(ffmpeg_cmd + '  >  /dev/null')
-            print('Compressed\t', file_name)
+            if file_ext in self.not_down:
+                print('Moving ', file_name)
+                os.system('cp -r ' + self.valid_unix_name(file) + ' ' + saveas)
+            
+            elif file_ext in self.video:
+                ffmpeg_cmd = "ffmpeg -i " + self.valid_unix_name(file) + "\
+                        -b:a 64k -ac 1 -vf scale=\"'w=-2:h=trunc(min(ih," + str(self.res) + ")/2)*2'\" \
+                        -crf 32 -profile:v baseline -level 3.0 -preset slow -v error -strict -2 -stats \
+                        -y -r 20 " + saveas
+                print('compressing\t', file_name)
+                os.system(ffmpeg_cmd + '  >  /dev/null')
+                print('Compressed\t', file_name)
 
-        elif file_ext not in self.not_down:
-            print('Moving ', file_name)
-            os.system('cp -r ' + self.valid_unix_name(file) + ' ' + saveas)
-        else:
-            print('File exists :  ', file_name)
+            else:
+                print('copying file :  ', file_name)
+                os.system('cp -r ' + self.valid_unix_name(file) + ' ' + saveas)
+
 
     def get_file(self, folder):
         os.chdir(folder)
         for file in os.listdir(folder):
             if not file.startswith('.'):
                 new_file = os.path.join(folder, file)
-                if os.path.isfile(new_file) and new_file.split('.')[-1] in self.video:
-                    self.files.append(new_file)
+                if os.path.isfile(new_file):
+                    if new_file.split('.')[-1] in self.video:
+                        self.files.append(new_file)
                 else:
                     self.get_file(new_file)	
 
@@ -117,4 +123,3 @@ local  :- To Where | gdrive url
         pool.map(self.compress, self.files)
         print("Done")
 
-Compress('/workspace/TIME Concept of Data Interpretation', '/home/gitpod')
